cnn model/requirements.txt
tensorflow==2.16.1
tf-keras==2.16.0
numpy<2.0.0
streamlit>=1.28.0
Pillow>=9.0.0

# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6